mypage user, pet, summary dto 생성

+++ SsedamPet_back/src/main/java/com/korit/ssedampet_back/mapper/MypageMapper.java
@@ -1,6 +1,5 @@
 package com.korit.ssedampet_back.mapper;
 
-import com.korit.ssedampet_back.dto.response.MypageRespDto;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
@@ -8,11 +7,11 @@ import java.util.List;
 
 @Mapper
 public interface MypageMapper {
-    MypageRespDto.UserDto findMypageUser(@Param("userId") Long userId);
+    MypageRespDto.UserDto findMypageUser(@Param("userId") int userId);
 
-    int countMyPosts(@Param("userId") Long userId);
+    int countMyPosts(@Param("userId") int userId);
 
-    int countMyLikedPosts(@Param("userId") Long userId);
+    int countMyLikedPosts(@Param("userId") int userId);
 
-    List<MypageRespDto.PetDto> findMyPets(@Param("userId") Long userId);
+    List<MypageRespDto.PetDto> findMyPets(@Param("userId") int userId);
 }
